arch/lapic: x2APIC support (MSR mode) + delete force-xAPIC panic
The kernel ran xAPIC-only and PANICKED when firmware locked the
CPU into x2APIC (EXTD set) — a configuration many server BIOSes
ship by default, so the kernel could not boot on that hardware at
all. Add real x2APIC support and delete the panic.

LapicInit now selects x2APIC (MSR) whenever CPUID.1:ECX[21]
advertises it (standard OS behavior; the only mode that works
when firmware already enabled it), else legacy xAPIC (MMIO).
LapicRead/LapicWrite dispatch on a one-time mode flag (x2APIC reg
= MSR 0x800 + (off>>4)); the four ICR sites (SmpSendIpi, the two
NMI broadcasts, the all-but-self fixed broadcast) route through a
single mode-aware LapicSendIcr — xAPIC writes ICR-hi/lo and spins
on delivery-status, x2APIC issues one wrmsr(0x830) with no poll.
WaitForIcrDelivery folded in and deleted. SmpSendIpi now takes a
full 32-bit APIC ID (no <<24 / static_cast<u8> truncation);
LapicCurrentId() normalises the ID read across modes and replaces
the three xAPIC-only `>>24` sites (percpu, smp BSP id, IOAPIC
route — the last keeps the low 8 bits since IOAPIC RTE dest is
8-bit in both modes). LapicIsReady is now a mode-independent flag
(g_lapic_mmio is always null under x2APIC).

Self-test apic-mode-selftest (Phase::Userland) asserts CPUID and
mode agree and the ID round-trips; PASSes on every guest. GAP
documented: legacy MADT LAPIC records are 8-bit, so >255 APIC IDs
need x2APIC MADT (type 9) parsing — a separate follow-on; fine
for all current targets. Roadmap updated; the cluster-broadcast
IPI's x2APIC-enablement prerequisite is now satisfied.

Build clean + ctest green; single-CPU boot verification in
progress (boot-critical path — timer/EOI/IPI now MSR).

+++ kernel/arch/x86_64/smp.h
@@ -60,16 +60,16 @@ u64 SmpStartAps();
 u64 SmpCpusOnline();
 
 /// Send an arbitrary IPI via the LAPIC Interrupt Command Register.
-/// `target_apic_id` is the destination LAPIC ID (bits 24..31 of the
-/// ICR high half). `icr_low` carries the delivery mode + vector +
-/// level/trigger bits per Intel SDM Vol. 3A "LAPIC Interrupt Command
-/// Register." Blocks until the delivery-status bit clears; panics if
-/// it stays pending for ~1e6 spin iterations (indicates a broken
-/// LAPIC or a CPU that never accepted the IPI).
+/// `target_apic_id` is the full 32-bit destination APIC ID (no
+/// 8-bit truncation — required for x2APIC). `icr_low` carries the
+/// delivery mode + vector + level/trigger bits per Intel SDM Vol.
+/// 3A. Routes through `arch::LapicSendIcr`, which is mode-aware
+/// (xAPIC: ICR-hi/lo + bounded delivery-status spin; x2APIC: one
+/// `wrmsr(0x830)`, no poll).
 ///
-/// Exposed now so future callers (TLB shootdown, resched-IPI, AP
-/// wake-up) share the same ICR dance rather than reimplementing it.
-void SmpSendIpi(u8 target_apic_id, u32 icr_low);
+/// Shared by TLB shootdown, resched-IPI, and AP wake-up so they
+/// don't reimplement the ICR dance.
+void SmpSendIpi(u32 target_apic_id, u32 icr_low);
 
 /// Reschedule-IPI vector. Set by the wake path on a remote CPU's
 /// runqueue to prompt that CPU to call Schedule() promptly rather
